+++ src/main/java/org/msgpack/rpc/loop/netty/MessageHandler.java
@@ -1,53 +1,28 @@
-////
-//// MessagePack-RPC for Java
-////
-//// Copyright (C) 2010 FURUHASHI Sadayuki
-////
-////    Licensed under the Apache License, Version 2.0 (the "License");
-////    you may not use this file except in compliance with the License.
-////    You may obtain a copy of the License at
-////
-////        http://www.apache.org/licenses/LICENSE-2.0
-////
-////    Unless required by applicable law or agreed to in writing, software
-////    distributed under the License is distributed on an "AS IS" BASIS,
-////    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-////    See the License for the specific language governing permissions and
-////    limitations under the License.
-////
-//package org.msgpack.rpc.loop.netty;
-//
-//import org.jboss.netty.channel.ChannelHandlerContext;
-//import org.jboss.netty.channel.ChannelStateEvent;
-//import org.jboss.netty.channel.MessageEvent;
-//import org.jboss.netty.channel.SimpleChannelUpstreamHandler;
-//import org.msgpack.rpc.transport.RpcMessageHandler;
-//import org.msgpack.type.Value;
-//
-//class MessageHandler extends SimpleChannelUpstreamHandler {
-//    private RpcMessageHandler handler;
-//    private ChannelAdaptor adaptor;
-//
-//    MessageHandler(RpcMessageHandler handler) {
-//        this.handler = handler;
-//    }
-//
-//    @Override
-//    public void channelOpen(ChannelHandlerContext ctx, ChannelStateEvent e)
-//            throws Exception {
-//        this.adaptor = new ChannelAdaptor(e.getChannel());
-//        ctx.sendUpstream(e);
-//    }
-//
-//    @Override
-//    public void messageReceived(ChannelHandlerContext ctx, MessageEvent e) {
-//        Object m = e.getMessage();
-//        if (!(m instanceof Value)) {
-//            ctx.sendUpstream(e);
-//            return;
-//        }
-//
-//        Value msg = (Value) m;
-//        handler.handleMessage(adaptor, msg);
-//    }
-//}
+package org.msgpack.rpc.loop.netty;
+
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelInboundHandlerAdapter;
+import org.msgpack.rpc.transport.RpcMessageHandler;
+import org.msgpack.type.Value;
+
+class MessageHandler extends ChannelInboundHandlerAdapter {
+
+    private final RpcMessageHandler _rpcHandler;
+
+    public MessageHandler(final RpcMessageHandler rpcHandler){
+        _rpcHandler = rpcHandler;
+    }
+
+    @Override
+    public void channelRead(ChannelHandlerContext ctx, Object msg) {
+
+        _rpcHandler.handleMessage(new ChannelAdaptor(ctx.channel()), (Value) msg);
+    }
+
+    @Override
+    public void exceptionCaught(final ChannelHandlerContext ctx, Throwable cause) throws Exception {
+
+        cause.printStackTrace();
+        ctx.close();
+    }
+}
