+++ src/main/java/org/msgpack/rpc/loop/netty/NettyTcpClientTransport.java
@@ -17,35 +17,20 @@
 //
 package org.msgpack.rpc.loop.netty;
 
-import java.io.EOFException;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
-import java.util.List;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
 import io.netty.bootstrap.Bootstrap;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufOutputStream;
-import io.netty.buffer.UnpooledByteBufAllocator;
 import io.netty.channel.*;
 import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioSocketChannel;
-import io.netty.handler.codec.ByteToMessageDecoder;
-import io.netty.handler.codec.MessageToByteEncoder;
-import org.msgpack.MessagePack;
-import org.msgpack.MessagePackable;
 import org.msgpack.rpc.Session;
 import org.msgpack.rpc.config.TcpClientConfig;
 import org.msgpack.rpc.transport.ClientTransport;
 import org.msgpack.rpc.transport.RpcMessageHandler;
-import org.msgpack.rpc.transport.PooledStreamClientTransport;
-import org.msgpack.type.Value;
-import org.msgpack.unpacker.Unpacker;
 
 class NettyTcpClientTransport implements ClientTransport {
-
-    private ByteBufOutputStream _bufOutput;
 
     private final Session _session;
     private final Bootstrap bootstrap;
@@ -123,111 +108,5 @@ class NettyTcpClientTransport implements ClientTransport {
                 _channels.offer(future.channel());
             }
         });
-    }
-
-    static class MessagePackDecoder extends ByteToMessageDecoder {
-
-        private final MessagePack _msgpack;
-
-        public MessagePackDecoder(final MessagePack msgpack){
-            _msgpack = msgpack;
-        }
-
-        @Override
-        protected void decode(final ChannelHandlerContext channelHandlerContext,
-                              final ByteBuf byteBuf,
-                              final List<Object> out) throws Exception {
-
-            //System.out.printf("[client transport] decode got bytebuf\n");
-
-            final ByteBuffer buffer = byteBuf.markReaderIndex().nioBuffer().slice();
-
-            try{
-                Unpacker unpacker = _msgpack.createBufferUnpacker(buffer);
-                out.add(unpacker.readValue());
-
-                //System.out.printf("[client transport] decode done\n");
-
-                byteBuf.skipBytes(buffer.position());
-            }
-            catch( EOFException e ){
-
-                //System.out.println("[client transport] not enough bytebuf");
-
-                byteBuf.resetReaderIndex();
-            }
-        }
-    }
-
-    static class MessagePackEncoder extends MessageToByteEncoder<Value> {
-
-        private final MessagePack _msgpack;
-
-        public MessagePackEncoder(final MessagePack msgpack){
-            _msgpack = msgpack;
-        }
-
-        @Override
-        protected void encode(ChannelHandlerContext ctx, Value msg, ByteBuf out) throws Exception {
-
-            //System.out.println("[client transport] encoding msg of Value");
-
-            _msgpack.createPacker(new ByteBufOutputStream(out)).write(msg);
-
-            ctx.flush();
-        }
-    }
-
-    static class MessagePackableEncoder extends MessageToByteEncoder<MessagePackable> {
-
-        private final MessagePack _msgpack;
-
-        public MessagePackableEncoder(final MessagePack msgpack){
-            _msgpack = msgpack;
-        }
-
-        @Override
-        protected void encode(ChannelHandlerContext ctx, MessagePackable msg, ByteBuf out) throws Exception {
-
-            //System.out.println("[client transport] encoding msg of Value from packable encoder");
-
-            msg.writeTo(_msgpack.createPacker(new ByteBufOutputStream(out)));
-        }
-
-        @Override
-        public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise) throws Exception {
-
-            super.write(ctx, msg, promise);
-
-            ctx.flush();
-
-            //System.out.println("[server transport] encoding msg of Value from packable encoder [flushed]");
-        }
-    }
-
-    static class MessageHandler extends ChannelInboundHandlerAdapter {
-
-        private final RpcMessageHandler _rpcHandler;
-
-        public MessageHandler(final RpcMessageHandler rpcHandler){
-            _rpcHandler = rpcHandler;
-        }
-
-        @Override
-        public void channelRead(ChannelHandlerContext ctx, Object msg) {
-
-            //System.out.printf("[client transport] message handler got msg: " + msg.getClass().getName());
-
-            Value value = (Value) msg;
-
-            _rpcHandler.handleMessage(new ChannelAdaptor(ctx.channel()), value);
-        }
-
-        @Override
-        public void exceptionCaught(final ChannelHandlerContext ctx, Throwable cause) throws Exception {
-
-            cause.printStackTrace();
-            ctx.close();
-        }
     }
 }
