--- src/main/java/org/msgpack/rpc/loop/netty/StreamPipelineFactory.java
@@ -1,42 +1,0 @@
-////
-//// MessagePack-RPC for Java
-////
-//// Copyright (C) 2010 FURUHASHI Sadayuki
-////
-////    Licensed under the Apache License, Version 2.0 (the "License");
-////    you may not use this file except in compliance with the License.
-////    You may obtain a copy of the License at
-////
-////        http://www.apache.org/licenses/LICENSE-2.0
-////
-////    Unless required by applicable law or agreed to in writing, software
-////    distributed under the License is distributed on an "AS IS" BASIS,
-////    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-////    See the License for the specific language governing permissions and
-////    limitations under the License.
-////
-//package org.msgpack.rpc.loop.netty;
-//
-//import org.jboss.netty.channel.Channels;
-//import org.jboss.netty.channel.ChannelPipeline;
-//import org.jboss.netty.channel.ChannelPipelineFactory;
-//import org.msgpack.MessagePack;
-//import org.msgpack.rpc.transport.RpcMessageHandler;
-//
-//class StreamPipelineFactory implements ChannelPipelineFactory {
-//    private RpcMessageHandler handler;
-//    private MessagePack messagePack;
-//
-//    StreamPipelineFactory(MessagePack messagePack, RpcMessageHandler handler) {
-//        this.handler = handler;
-//        this.messagePack = messagePack;
-//    }
-//
-//    public ChannelPipeline getPipeline() throws Exception {
-//        ChannelPipeline p = Channels.pipeline();
-//        p.addLast("msgpack-decode-stream", new MessagePackStreamDecoder(messagePack));
-//        p.addLast("msgpack-encode", new MessagePackEncoder(messagePack));
-//        p.addLast("message", new MessageHandler(handler));
-//        return p;
-//    }
-//}
